pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bilibili-extractor"
version = "1.0.0"
description = "Extract text from Bilibili videos using subtitles, ASR, and OCR"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Bilibili Extractor Team"}
]
keywords = ["bilibili", "video", "text-extraction", "asr", "subtitle"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "ffmpeg-python>=0.2.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
asr = [
    "funasr>=1.0.0",
    "openai-whisper>=20230314",
]
ocr = [
    "paddleocr>=2.7.0",
    "opencv-python>=4.8.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.82.0",
    "black>=23.7.0",
    "mypy>=1.4.0",
]

[project.scripts]
bilibili-extractor = "bilibili_extractor.__main__:main"

[project.urls]
Homepage = "https://github.com/bilibili-extractor/bilibili-extractor"
Repository = "https://github.com/bilibili-extractor/bilibili-extractor"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
